+++ source/blender/editors/interface/interface_intern.hh
@@ -936,7 +936,7 @@ const char *button_placeholder_get(Button *but);
  */
 std::optional<StringRef> button_edit_unit_hint_get(const Button &but);
 
-void def_but_icon(Button *but, int icon, int flag);
+void def_but_icon(Button *but, int icon, int64_t flag);
 /**
  * Avoid using this where possible since it's better not to ask for an icon in the first place.
  */
@@ -1472,7 +1472,7 @@ void draw_menu_item(const uiFontStyle *fstyle,
                     bool use_unpadded,
                     const char *name,
                     int iconid,
-                    int but_flag,
+                    int64_t but_flag,
                     MenuItemSeparatorType separator_type,
                     int *r_xmax);
 void draw_preview_item(const uiFontStyle *fstyle,
@@ -1480,7 +1480,7 @@ void draw_preview_item(const uiFontStyle *fstyle,
                        float zoom,
                        const char *name,
                        int iconid,
-                       int but_flag,
+                       int64_t but_flag,
                        FontStyleAlign text_align);
 /**
  * Version of #draw_preview_item() that does not draw the menu background and item text based on
